Added some opcode support and a dump function
I added support for a few opcodes (up through 8xy4), along with a python function that dumps register and memory values. The dump function is more for me to use when debugging, but I added it , so it needs to go here. Note that declarations for the opcode handlers I added haven't been added to chip8lib.h yet, and they also aren't being called by the global handler.